+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/domain/cadastro/TamanhoRoupa.kt
@@ -5,13 +5,13 @@ import jakarta.validation.constraints.Max
 
 @Entity
 class TamanhoRoupa(
-    @field:Id @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idTamanhoRoupa:Int = 0,
+    @field:Id @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idTamanhoRoupa:Long = 0,
     @field:Max(10) private var tamanho:String
 ) {
-    fun getId():Int{
+    fun getId():Long{
         return idTamanhoRoupa
     }
-    fun setId(novoId:Int){
+    fun setId(novoId:Long){
         idTamanhoRoupa = novoId
     }
     fun getTamanho():String{
