+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/domain/doacao/AssuntoRequisicao.kt
@@ -10,14 +10,14 @@ import jakarta.validation.constraints.NotBlank
 @Entity
 class AssuntoRequisicao(
     @field:Id
-    @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idAssuntoRequisicao: Int = 0,
+    @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idAssuntoRequisicao: Long = 0,
     @field:NotBlank @field:Max(150) private var assunto:String
 ) {
 
-    fun getId():Int{
+    fun getId():Long{
         return idAssuntoRequisicao
     }
-    fun setId(novoId:Int){
+    fun setId(novoId:Long){
         idAssuntoRequisicao = novoId
     }
     fun getAssunto():String{
