+++ projeto-tfg/src/main/kotlin/school/sptech/projetotfg/domain/gerenciamento/Permissao.kt
@@ -5,13 +5,13 @@ import jakarta.validation.constraints.Max
 
 @Entity
 class Permissao(
-    @field:Id @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idPermissao:Int = 0,
+    @field:Id @field:GeneratedValue(strategy = GenerationType.IDENTITY) private var idPermissao:Long = 0,
     @field:Max(30) private var permissao:String
 ) {
-    fun getId():Int{
+    fun getId():Long{
         return idPermissao
     }
-    fun setId(novoId:Int){
+    fun setId(novoId:Long){
         idPermissao = novoId
     }
     fun getPermissao():String{
